chore: move config watcher to graph command
