+++ src/commands/base-command.js
@@ -1,5 +1,4 @@
 // @ts-check
-const events = require('events')
 const process = require('process')
 const { format } = require('util')
 
@@ -29,7 +28,6 @@ const {
   pollForToken,
   sortOptions,
   track,
-  watchDebounced,
 } = require('../utils')
 
 // Netlify CLI client id. Lives in bot@netlify.com
@@ -120,11 +118,6 @@ class BaseCommand extends Command {
           // @ts-ignore cannot type actionCommand as BaseCommand
           await this.init(actionCommand)
           debug(`${name}:preAction`)('end')
-        })
-        .hook('postAction', async () => {
-          if (this.configWatcherHandle) {
-            await this.configWatcherHandle.close()
-          }
         })
     )
   }
@@ -437,27 +430,10 @@ class BaseCommand extends Command {
     const globalConfig = await getGlobalConfig()
     const { NetlifyAPI } = await jsClient
 
-    const configWatcher = new events.EventEmitter()
-
-    // Only set up a watcher if we know the config path.
-    if (configPath) {
-      const configWatcherHandle = await watchDebounced(configPath, {
-        depth: 1,
-        onChange: async () => {
-          const { config: newConfig } = await actionCommand.getConfig({ cwd, state, token, ...apiUrlOpts })
-
-          const normalizedNewConfig = normalizeConfig(newConfig)
-          configWatcher.emit('change', normalizedNewConfig)
-        },
-      })
-
-      // chokidar handler
-      this.configWatcherHandle = configWatcherHandle
-    }
-
     actionCommand.netlify = {
       // api methods
       api: new NetlifyAPI(token || '', apiOpts),
+      apiOpts,
       repositoryRoot,
       // current site context
       site: {
@@ -480,8 +456,6 @@ class BaseCommand extends Command {
       globalConfig,
       // state of current site dir
       state,
-      // netlify.toml file watcher
-      configWatcher,
     }
     debug(`${this.name()}:init`)('end')
   }
